Fixed some compilation warnings
git-svn-id: https://ufo2000.svn.sourceforge.net/svnroot/ufo2000/trunk@885 80a73507-52ba-0310-bdc1-b9b4a1856fab

+++ src/global.h
@@ -312,9 +312,9 @@ extern int g_pause;
 extern int g_game_receiving;
 extern int g_fast_forward;
 // id of a last received or sended packet to use in the debug info.
-extern long int g_current_packet_num;
+extern int g_current_packet_num;
 //position of player who send currently processed packet to use in the debug info.
-extern long int g_current_packet_pos;
+extern int g_current_packet_pos;
 
 extern int debug_save_state_sender;
 extern int debug_save_state_id;
